Working Chassis Code
This Chassis code works! Due to how the acceleration thingy works, that is NOT in the code. That is done through the REV Robotics Spark Max Client, so make sure that they are properly set up. (Mine was set to 0.4 seconds to go from 0 to 100%, which I believe was in Gianluca's safe motor speed range. Will Need a chassis to test that for sure)