Working Chassis Code (#36)
* Phoenix Installed and built

Installed Phoenix and built it!

* Implemented Chassis into subsystem folder

* Added REVRobotics implementation

* Implemented Oblog, added potential logging of speed controllers

Added the implementation necessary to use Oblog, and attempted to write code that may log data into Shuffleboard.

* Editting Chassis code to make it work better

Needs Joystick control.

* Working on making working code for motors

* Working motors!!!!

* Consolidation

Taking out the SparkMaxs from the robot class and shoving them into the chassis class.

* added extra motors for chassis, started button tapped code

Added two extra motors for the chassis motors. Started putting in code to make a toggle for the mode changer.

* Added drive train chooser and PID

Added the Drive Train selector into the robot.java allowing the drive train type to be chosen before the robot is enabled. Also, added PID thingies to the sparkMax motors, but I have no idea if they work or how they work.

* Added stuff

* Added Motor Widgets, removed @logs and @configs from code

Added motor widgets the old school way and removed the Oblog logging thingies. I don't know how to use Oblog, and I understand this way at least a little bit, so plz don't bite me or kill me in my sleep

* Did things, made changes

Added a differential drive widget (maybe) and set the default value for the motor widgets to zero instead of one (to avoid the motors going full send on some poor unsuspecting person)

* Working Drive Code

This includes working drive code for Tank, arcade, and curvature drives! It also has Shuffleboard control of the motors individually and in their controller groups, allowing for better testing of the motors. All control done through Shuffleboard is done by old school widget control, so DO NOT DELETE THE WIDGET CLASSES unless you can successfully re-create the work done by the widgets with Oblog. Since I don't know how to use Oblog, I used the old fashioned way.

* Working Chassis Code!!!!

This is working chassis code!!!! Due to an unfamiliarity with Oblog, I am still NOT USING Oblog for my widgets... I DID ADD the neccessary code to use Oblog, but have decided to comment it out for now for my own sake... this is viable to change in the future, just be made aware that this is how it is for now.

* Working Chassis Code

This Chassis code works! Due to how the acceleration thingy works, that is NOT in the code. That is done through the REV Robotics Spark Max Client, so make sure that they are properly set up. (Mine was set to 0.4 seconds to go from 0 to 100%, which I believe was in Gianluca's safe motor speed range. Will Need a chassis to test that for sure)