Velocity control work (#88)
* Velocity Work!!!!

Making all of this stuff to try to make a working velocity control mode for our drive train... it's a lot of work, but it just might work! Needs more work... and sugar.

* Created the velo city stuffs!

Created the controls for the city of velos, and hopefully, it will work as intended... maybe.

* Added Scott Drive

Added Scott Drive, which is essentially the specially designed drive system requested by our ingenius driver, Scott Ulatowski!!!!

* Added deadzone for controller.

Added a deadzone for the controller, so that if you lightly tap the bumper it might not drive the robot accidentally. Current dead zone is 5%, but that can be adjusted as needed.

* Added comment to a shuffleboard tab

Juse wanted to let it be knonw that the tab isn't abandoned, it just doesn't realize when it is being used.

* Removed unused imports

:)

* Added current limit

:)

* Fixed the kp's

* added i max and izone

* Changed current limit

Made current limit 40 instead of 5... apparently 5 isn't very good for the robot and makes it squeaky...

* Lower P value (for safety reasons)

and because Sir Rylan the Wise said to...

* MAking changes for working velocity control mode!

* Adding testing point thingies!

* PIDF stuff done

Worked on the PIDF stuff so as to get the robot to actually drive properly... has it worked? Kind of... the accuracy of this is still TBD.

* Potential fix for PIDF Loop

Changed the squared Inputs option, so that it is now controlled by us and not on all the time, giving us better values to work with and test with.

* Made adjustments to PID Control

Made it so that PID stuff can only be changed if we have it set to true.

Co-authored-by: Rylan Meilutis <44624900+Rylan-Meilutis@users.noreply.github.com>